requirements.txt
torch>=1.11.0
numpy>=1.20.3
pandas==1.5.3
scipy>=1.7.3
scikit-learn>=1.0.2
wilds
ftfy
regex
git+https://github.com/openai/CLIP.git
open_clip_torch